requirements.txt
strands-agents
strands-agents-tools
python-dotenv
